+++ controler/DaemonControler.php
@@ -45,7 +45,10 @@ class DaemonControler extends PastellControler
      */
     public function verrouAction(): void
     {
-        $this->verifDroit(0, DroitService::getDroitLecture(DroitService::DROIT_DAEMON));
+        $this->verifDroit(
+            EntiteSQL::ID_E_ENTITE_RACINE,
+            DroitService::getDroitLecture(DroitService::DROIT_DAEMON)
+        );
         $this->setViewParameter('job_queue_info_list', $this->getJobQueueSQL()->getCountJobByVerrouAndEtat());
         $this->setViewParameter('menu_gauche_select', 'Daemon/verrou');
         $this->setViewParameter('template_milieu', 'DaemonVerrou');
@@ -74,28 +77,18 @@ class DaemonControler extends PastellControler
      */
     private function indexData(): void
     {
-        $this->verifDroit(0, DroitService::getDroitLecture(DroitService::DROIT_DAEMON));
-        $this->setDroitsDaemon(0);
-        $this->setViewParameter('nb_total_workers', $this->getDaemonSQL()->getNbTotalWorkers());
+        $this->verifDroit(
+            EntiteSQL::ID_E_ENTITE_RACINE,
+            DroitService::getDroitLecture(DroitService::DROIT_DAEMON)
+        );
+        $this->setDroitsDaemon(EntiteSQL::ID_E_ENTITE_RACINE);
         $this->setViewParameter('nb_worker_actif', $this->getWorkerSQL()->getNbActif());
-        $this->setViewParameter('nb_total_workers', $this->getConfigurationSQL()->getConfiguration(ConfigurationSQL::NB_WORKERS));
+        $this->setViewParameter('nb_workers', $this->getDaemonSQL()->getNbWorkers());
         $this->setViewParameter('job_stat_info', $this->getJobQueueSQL()->getStatInfo());
         $this->setViewParameter('daemon_pid', $this->getDaemonManager()->getDaemonPID());
         $this->setViewParameter('sub_title', 'Liste de tous les travaux');
         $this->setViewParameter('return_url', urlencode('Daemon/index'));
-
-        $job_list = $this->getJobQueueSQL()->getAllJobs();
-        foreach ($job_list as $job) {
-            /** @var Job $job */
-            $job->daemon =  $this->getDaemonSQL()->getDaemon($job->id_daemon);
-
-            $worker = $this->getWorkerSQL()->getWorker($job->id_job);
-            if ($worker !== null) {
-                $job->worker = $worker;
-            }
-        }
-        $this->setViewParameter('job_list', $job_list);
-        $this->setViewParameter('daemon', $this->getDaemonSQL()->getDaemon(0));
+        $this->setViewParameter('job_list', $this->getJobQueueSQL()->getAllJobs());
     }
 
     /**
@@ -104,7 +97,10 @@ class DaemonControler extends PastellControler
      */
     public function globalDaemonStartAction(): void
     {
-        $this->verifDroit(0, DroitService::getDroitEdition(DroitService::DROIT_DAEMON));
+        $this->verifDroit(
+            EntiteSQL::ID_E_ENTITE_RACINE,
+            DroitService::getDroitEdition(DroitService::DROIT_DAEMON)
+        );
         try {
             $this->getDaemonManager()->start();
             $this->getLogger()->info('Daemon start manually');
@@ -131,9 +127,12 @@ class DaemonControler extends PastellControler
      */
     public function globalDaemonStopAction(): void
     {
-        $this->verifDroit(0, DroitService::getDroitEdition(DroitService::DROIT_DAEMON));
+        $this->verifDroit(
+            EntiteSQL::ID_E_ENTITE_RACINE,
+            DroitService::getDroitEdition(DroitService::DROIT_DAEMON)
+        );
         $this->getDaemonManager()->stop();
-        if ($this->getDaemonManager()->status() == DaemonManager::IS_STOPPED) {
+        if ($this->getDaemonManager()->status() === DaemonManager::IS_STOPPED) {
             $this->setLastMessage('Les gestionnaires de tâches ont été arrêtés');
         } else {
             $this->setLastError("Une erreur s'est produite lors de la tentative d'arrêt des gestionnaires de tâches");
@@ -163,6 +162,10 @@ class DaemonControler extends PastellControler
                 $this->setLastMessage('Le travail a été suspendu');
             }
         } elseif ($id_verrou || $etat_source || $etat_cible) {
+            $this->verifDroit(
+                EntiteSQL::ID_E_ENTITE_RACINE,
+                DroitService::getDroitEdition(DroitService::DROIT_DAEMON)
+            );
             $this->getJobQueueSQL()->lockByVerrouAndEtat($id_verrou, $etat_source, $etat_cible);
             $this->setLastMessage('Les travaux correspondants ont été suspendus');
         } else {
@@ -196,6 +199,10 @@ class DaemonControler extends PastellControler
                 $this->setLastMessage('Le travail a été réactivé');
             }
         } elseif ($id_verrou || $etat_source || $etat_cible) {
+            $this->verifDroit(
+                EntiteSQL::ID_E_ENTITE_RACINE,
+                DroitService::getDroitEdition(DroitService::DROIT_DAEMON)
+            );
             $this->getJobQueueSQL()->unlockByVerrouAndEtat($id_verrou, $etat_source, $etat_cible);
             $this->setLastMessage('Les travaux correspondants ont été réactivés');
         } else {
@@ -213,7 +220,10 @@ class DaemonControler extends PastellControler
     {
         $this->getWorkerSQL()->menageAll();
         $this->getJobQueueSQL()->unlockAll();
-        $this->verifDroit(0, DroitService::getDroitEdition(DroitService::DROIT_DAEMON));
+        $this->verifDroit(
+            EntiteSQL::ID_E_ENTITE_RACINE,
+            DroitService::getDroitEdition(DroitService::DROIT_DAEMON)
+        );
         $this->redirect('Daemon/index');
     }
 
@@ -239,7 +249,7 @@ class DaemonControler extends PastellControler
         }
 
         $this->verifDroit($job->id_e, DroitService::getDroitEdition(DroitService::DROIT_DAEMON));
-        $this->getJobQueueSQL()->lock($id_job);
+        $this->getJobQueueSQL()->lock($job->id_job);
 
         $process = new Process(['kill', '-9', $worker->pid]);
         $process->run();
@@ -264,7 +274,10 @@ class DaemonControler extends PastellControler
         $recuperateur = $this->getGetInfo();
         $this->setViewParameter('menu_gauche_select', 'Daemon/job');
 
-        $this->verifDroit(0, DroitService::getDroitEdition(DroitService::DROIT_DAEMON));
+        $this->verifDroit(
+            EntiteSQL::ID_E_ENTITE_RACINE,
+            DroitService::getDroitEdition(DroitService::DROIT_DAEMON)
+        );
         $this->setViewParameter('twigTemplate', 'daemon/job.html.twig');
         $this->setViewParameter('page_title', 'Gestionnaire de tâches');
         $filtre = $recuperateur->get('filtre', '');
@@ -312,7 +325,10 @@ class DaemonControler extends PastellControler
      */
     public function detailAction(): void
     {
-        $this->verifDroit(0, DroitService::getDroitEdition(DroitService::DROIT_DAEMON));
+        $this->verifDroit(
+            EntiteSQL::ID_E_ENTITE_RACINE,
+            DroitService::getDroitEdition(DroitService::DROIT_DAEMON)
+        );
         $id_job = $this->getGetInfo()->get('id_job');
 
         $this->setViewParameter('page_title', "Détail du travail #{$id_job}");
@@ -329,13 +345,16 @@ class DaemonControler extends PastellControler
      * @throws NotFoundException
      * @throws LastErrorException
      */
-    public function frequenceConfigAction()
+    public function frequenceConfigurationAction()
     {
-        $this->verifDroit(0, DroitService::getDroitEdition(DroitService::DROIT_DAEMON));
+        $this->verifDroit(
+            EntiteSQL::ID_E_ENTITE_RACINE,
+            DroitService::getDroitEdition(DroitService::DROIT_DAEMON)
+        );
 
         $this->setViewParameter('page_title', 'Configuration de la fréquence des connecteurs');
-        $this->setViewParameter('template_milieu', 'DaemonFrequenceConfig');
-        $this->setViewParameter('menu_gauche_select', 'Daemon/frequenceConfig');
+        $this->setViewParameter('template_milieu', 'DaemonFrequenceConfiguration');
+        $this->setViewParameter('menu_gauche_select', 'Daemon/frequenceConfiguration');
         $this->setViewParameter('nouveau_bouton_url', ['Ajouter' => 'Daemon/editFrequence']);
         $this->setViewParameter('connecteur_frequence_list', $this->getConnecteurFrequenceSQL()->getAll());
         $this->renderDefault();
@@ -348,16 +367,21 @@ class DaemonControler extends PastellControler
      */
     public function editFrequenceAction(): void
     {
-        $this->verifDroit(0, DroitService::getDroitEdition(DroitService::DROIT_DAEMON));
+        $this->verifDroit(
+            EntiteSQL::ID_E_ENTITE_RACINE,
+            DroitService::getDroitEdition(DroitService::DROIT_DAEMON)
+        );
         $id_cf = $this->getGetInfo()->getInt('id_cf');
-        $connecteurFrequence = $this->getConnecteurFrequenceSQL()->getConnecteurFrequence($id_cf) ?: new ConnecteurFrequence();
+        $connecteurFrequence = $this->getConnecteurFrequenceSQL()->getConnecteurFrequence(
+            $id_cf
+        ) ?: new ConnecteurFrequence();
 
         $this->setViewParameter('connecteurFrequence', $connecteurFrequence);
 
         $verbe = $connecteurFrequence->id_cf ? 'Modification' : 'Ajout';
         $this->setViewParameter('page_title', "$verbe d'une fréquence de connecteur");
-        $this->setViewParameter('template_milieu', 'DaemmonEditFrequence');
-        $this->setViewParameter('menu_gauche_select', 'Daemon/frequenceConfig');
+        $this->setViewParameter('template_milieu', 'DaemonEditFrequence');
+        $this->setViewParameter('menu_gauche_select', 'Daemon/frequenceConfiguration');
         $this->renderDefault();
     }
 
@@ -422,21 +446,32 @@ class DaemonControler extends PastellControler
      */
     public function doEditFrequenceAction(): void
     {
-        $this->verifDroit(0, DroitService::getDroitEdition(DroitService::DROIT_DAEMON));
+        $this->verifDroit(
+            EntiteSQL::ID_E_ENTITE_RACINE,
+            DroitService::getDroitEdition(DroitService::DROIT_DAEMON)
+        );
         $connecteurFrequence = new ConnecteurFrequence($this->getPostInfo()->getAll());
         $id_cf = $this->getConnecteurFrequenceSQL()->edit($connecteurFrequence);
         $this->redirect("Daemon/connecteurFrequenceDetail?id_cf=$id_cf");
     }
 
-    public function connecteurFrequenceDetailAction()
+    /**
+     * @throws NotFoundException
+     * @throws LastMessageException
+     * @throws LastErrorException
+     */
+    public function connecteurFrequenceDetailAction(): void
     {
-        $this->verifDroit(0, DroitService::getDroitEdition(DroitService::DROIT_DAEMON));
+        $this->verifDroit(
+            EntiteSQL::ID_E_ENTITE_RACINE,
+            DroitService::getDroitEdition(DroitService::DROIT_DAEMON)
+        );
         $id_cf = $this->getGetInfo()->getInt('id_cf');
         $connecteurFrequence = $this->verifConnecteur($id_cf);
         $this->setViewParameter('connecteurFrequence', $connecteurFrequence);
         $this->setViewParameter('page_title', "Détail sur la fréquence d'un connecteur");
         $this->setViewParameter('template_milieu', 'DaemonFrequenceDetail');
-        $this->setViewParameter('menu_gauche_select', 'Daemon/frequenceConfig');
+        $this->setViewParameter('menu_gauche_select', 'Daemon/frequenceConfiguration');
         $this->renderDefault();
     }
 
@@ -446,12 +481,15 @@ class DaemonControler extends PastellControler
      */
     private function verifConnecteur($id_cf): ConnecteurFrequence
     {
-        $this->verifDroit(0, DroitService::getDroitEdition(DroitService::DROIT_DAEMON));
+        $this->verifDroit(
+            EntiteSQL::ID_E_ENTITE_RACINE,
+            DroitService::getDroitEdition(DroitService::DROIT_DAEMON)
+        );
         $connecteurFrequence = $this->getConnecteurFrequenceSQL()->getConnecteurFrequence($id_cf);
 
-        if (! $connecteurFrequence) {
+        if (!$connecteurFrequence) {
             $this->setLastError("Impossible de trouver le connecteur $id_cf");
-            $this->redirect('Daemon/frequenceConfig');
+            $this->redirect('Daemon/frequenceConfiguration');
         }
         return $connecteurFrequence;
     }
@@ -462,11 +500,14 @@ class DaemonControler extends PastellControler
      */
     public function deleteFrequenceAction(): void
     {
-        $this->verifDroit(0, DroitService::getDroitEdition(DroitService::DROIT_DAEMON));
+        $this->verifDroit(
+            EntiteSQL::ID_E_ENTITE_RACINE,
+            DroitService::getDroitEdition(DroitService::DROIT_DAEMON)
+        );
         $id_cf = $this->getGetInfo()->get('id_cf');
         $this->getConnecteurFrequenceSQL()->delete($id_cf);
         $this->setLastMessage('La fréquence a été supprimée');
-        $this->redirect('Daemon/frequenceConfig');
+        $this->redirect('Daemon/frequenceConfiguration');
     }
 
     /**
@@ -477,7 +518,6 @@ class DaemonControler extends PastellControler
     {
         $id_job = $this->getGetInfo()->get('id_job');
         $id_connecteur = $this->getGetInfo()->get('id_ce', 'Connecteur/index');
-
         if ($id_job) {
             $job = $this->getJobQueueSQL()->getJob($id_job);
             if ($job === null) {
@@ -490,7 +530,6 @@ class DaemonControler extends PastellControler
         } else {
             $this->setLastError('Identifiant de travail manquant');
         }
-
         $this->redirect("Connecteur/edition?id_ce=$id_connecteur");
     }
 
@@ -504,7 +543,6 @@ class DaemonControler extends PastellControler
         $id_job = $this->getGetInfo()->get('id_job');
         $id_document = $this->getGetInfo()->get('id_d');
         $id_e = $this->getGetInfo()->get('id_e');
-
         if ($id_job) {
             $job = $this->getJobQueueSQL()->getJob($id_job);
             if ($job === null) {
@@ -525,40 +563,31 @@ class DaemonControler extends PastellControler
      * @throws NotFoundException
      * @throws LastErrorException
      */
-    public function configAction(): void
+    public function configurationAction(): void
     {
+        $this->verifDroit(
+            EntiteSQL::ID_E_ENTITE_RACINE,
+            DroitService::getDroitEdition(DroitService::DROIT_DAEMON)
+        );
         $recuperateur = $this->getGetInfo();
         $offset = $recuperateur->getInt('offset', 0);
         $search = $recuperateur->get('search', '');
         $this->setViewParameter('search', $search);
         $this->setViewParameter('offset', $offset);
-        $entity_list = $this->getEntiteSQL()->getAllDaemonsInfo();
-        $global_daemon = $this->getDaemonSQL()->getGlobalDaemon();
-        array_unshift(
-            $entity_list,
-            [
-                'denomination' => 'Entité racine',
-                'is_active' => true,
-                'id_daemon' => $global_daemon->id_daemon,
-                'nb_workers' =>  $global_daemon->nb_workers,
-                'state' => $global_daemon->state,
-                'id_e' => $global_daemon->id_e,
-            ]
-        );
+        $entity_list = $this->getEntiteListe()->getAllDaemonsInfo($offset, $search);
         $this->setViewParameter('entity_list', $entity_list);
 
-        $nb_workers = $this->getDaemonSQL()->getNbTotalWorkers();
+        $nb_workers = $this->getDaemonSQL()->getNbWorkers();
         $nb_allocated_workers = $this->getDaemonSQL()->getNbAllocatedWorkers();
-        $nb_shared_workers =  $nb_workers - $nb_allocated_workers;
+        $nb_shared_workers = $nb_workers - $nb_allocated_workers;
 
-        $this->setViewParameter('nb_workers', $this->getDaemonManager()->getNbWorkers());
+        $this->setViewParameter('nb_workers', $nb_workers);
         $this->setViewParameter('nb_allocated_workers', $nb_allocated_workers);
         $this->setViewParameter('nb_shared_workers', $nb_shared_workers);
         $this->setViewParameter('global_daemon_status', $this->getDaemonManager()->status());
         $this->setViewParameter('page_title', 'Configuration des gestionnaires de tâches');
-        $this->setViewParameter('template_milieu', 'DaemonConfig');
-        $this->setViewParameter('menu_gauche_select', 'Daemon/config');
-        $this->setViewParameter('system_edition', $this->verifDroit(0, 'system:edition'));
+        $this->setViewParameter('template_milieu', 'DaemonConfiguration');
+        $this->setViewParameter('menu_gauche_select', 'Daemon/configuration');
         $this->renderDefault();
     }
 
@@ -567,13 +596,16 @@ class DaemonControler extends PastellControler
      * @throws NotFoundException
      * @throws LastErrorException
      */
-    public function editConfigAction(): void
+    public function editConfigurationAction(): void
     {
+        $this->verifDroit(
+            EntiteSQL::ID_E_ENTITE_RACINE,
+            DroitService::getDroitEdition(DroitService::DROIT_DAEMON)
+        );
+        $this->setViewParameter('nb_workers', $this->getDaemonSQL()->getNbWorkers());
         $this->setViewParameter('page_title', 'Configuration des gestionnaires de tâches');
-        $this->setViewParameter('template_milieu', 'DaemonEditConfig');
-        $this->setViewParameter('menu_gauche_select', 'Daemon/editConfig');
-        $this->setViewParameter('nb_workers', $this->getDaemonManager()->getNbWorkers());
-        $this->setViewParameter('system_edition', $this->verifDroit(0, 'system:edition'));
+        $this->setViewParameter('template_milieu', 'DaemonEditConfiguration');
+        $this->setViewParameter('menu_gauche_select', 'Daemon/configuration');
         $this->renderDefault();
     }
 
@@ -581,21 +613,25 @@ class DaemonControler extends PastellControler
      * @throws LastMessageException
      * @throws LastErrorException
      */
-    public function doEditConfigAction(): void
+    public function doEditConfigurationAction(): void
     {
-        $this->verifDroit(0, 'system:edition');
-        $nb_workers = (int) $this->getPostInfo()->get('nb_workers');
+        $this->verifDroit(
+            EntiteSQL::ID_E_ENTITE_RACINE,
+            DroitService::getDroitEdition(DroitService::DROIT_DAEMON)
+        );
+        $nb_workers = (int)$this->getPostInfo()->get('nb_workers');
         $allocatedWorkers = $this->getDaemonSQL()->getNbAllocatedWorkers();
         if ($allocatedWorkers > $nb_workers) {
-            $this->setLastError("Impossible de réduire le nombre de processus à $nb_workers, 
-            car $allocatedWorkers processus sont déjà alloués à des gestionnaires de tâches actifs. 
-            <br>Veuillez libérer des processus avant de réduire la configuration.");
+            $this->setLastError(
+                "Impossible de réduire le nombre de processus à $nb_workers 
+            car $allocatedWorkers processus sont déjà alloués à des gestionnaires de tâches actifs.
+            <br>Veuillez libérer des processus avant de réduire la configuration."
+            );
         } else {
             $this->getDaemonSQL()->setNbWorkers($nb_workers);
-            $this->getDaemonSQL()->refreshAvailableWorkers();
             $this->setLastMessage('La configuration des processus à été mise à jour');
         }
-        $this->redirect('Daemon/config');
+        $this->redirect('Daemon/configuration');
     }
 
     /**
@@ -604,22 +640,25 @@ class DaemonControler extends PastellControler
      */
     public function allocateAction(): void
     {
-        $this->verifDroit(0, 'system:edition');
+        $this->verifDroit(
+            EntiteSQL::ID_E_ENTITE_RACINE,
+            DroitService::getDroitEdition(DroitService::DROIT_DAEMON)
+        );
         $allocatedWorkers = $_POST['data'] ?? [];
         $nb_workers_to_allocate = array_sum($allocatedWorkers);
-        $nb_workers = $this->getDaemonSQL()->getNbTotalWorkers();
+        $nb_workers = $this->getDaemonSQL()->getNbWorkers();
         if ($nb_workers_to_allocate >= $nb_workers) {
             $this->setLastError('Au moins un processus partagé doit rester disponible.');
         } else {
-            foreach ($allocatedWorkers as $id_daemon => $nb_workers) {
-                if ($id_daemon <= 1) {
+            foreach ($allocatedWorkers as $id_daemon => $nb_workers_daemon) {
+                if ($id_daemon === DaemonSQL::GLOBAL_DAEMON) {
                     continue;
                 }
-                $this->getDaemonManager()->allocateWorkers($id_daemon, (int)$nb_workers);
+                $this->getDaemonManager()->allocateWorkers($id_daemon, (int)$nb_workers_daemon);
             }
             $this->setLastMessage('Processus alloués avec succès');
         }
-        $this->redirect('/Daemon/config');
+        $this->redirect('Daemon/configuration');
     }
 
     /**
@@ -629,20 +668,23 @@ class DaemonControler extends PastellControler
      */
     public function deleteDaemonAction(): void
     {
-        $this->verifDroit(0, 'system:edition');
+        $this->verifDroit(
+            EntiteSQL::ID_E_ENTITE_RACINE,
+            DroitService::getDroitEdition(DroitService::DROIT_DAEMON)
+        );
         $id_daemon = $this->getGetInfo()->getInt('id_daemon');
-        if ($id_daemon === 1) {
+        if ($id_daemon === DaemonSQL::GLOBAL_DAEMON) {
             $this->setLastError('Impossible de supprimer le gestionnaire de tâches global');
-            $this->redirect('Daemon/config');
+            $this->redirect('Daemon/configuration');
         }
         $daemon = $this->getDaemonSQL()->getDaemon($id_daemon);
         if ($daemon === null) {
             $this->setLastError('Impossible de trouver le gestionnaire de tâches');
-            $this->redirect('Daemon/config');
+            $this->redirect('Daemon/configuration');
         }
         $this->setViewParameter('page_title', 'Suppression du gestionnaire de tâches');
         $this->setViewParameter('template_milieu', 'DaemonDelete');
-        $this->setViewParameter('menu_gauche_select', 'Daemon/editConfig');
+        $this->setViewParameter('menu_gauche_select', 'Daemon/editConfiguration');
         $this->setViewParameter('daemon', $daemon);
         $this->setViewParameter('entite', $this->getEntiteSQL()->getInfo($daemon->id_e));
         $this->renderDefault();
@@ -655,16 +697,19 @@ class DaemonControler extends PastellControler
      */
     public function doDeleteDaemonAction(): void
     {
-        $this->verifDroit(0, 'system:edition');
+        $this->verifDroit(
+            EntiteSQL::ID_E_ENTITE_RACINE,
+            DroitService::getDroitEdition(DroitService::DROIT_DAEMON)
+        );
         $id_daemon = $this->getGetInfo()->getInt('id_daemon');
         $daemon = $this->getDaemonSQL()->getDaemon($id_daemon);
         if ($daemon === null) {
             $this->setLastError('Impossible de trouver le gestionnaire de tâches');
-            $this->redirect('Daemon/config');
+            $this->redirect('Daemon/configuration');
         }
         $this->getDaemonManager()->removeDaemon($daemon->id_daemon);
         $this->setLastMessage('Gestionnaire de tâches supprimé avec succès');
-        $this->redirect('Daemon/config');
+        $this->redirect('Daemon/configuration');
     }
 
     /**
@@ -675,14 +720,17 @@ class DaemonControler extends PastellControler
     public function daemonStartAction(): void
     {
         $id_daemon = $this->getGetInfo()->getInt('id_daemon');
-        $return_url = $this->getGetInfo()->get('return_url', 'Daemon/config');
+        $return_url = $this->getGetInfo()->get('return_url', 'Daemon/configuration');
         $daemon = $this->getDaemonSQL()->getDaemon($id_daemon);
         if ($daemon === null) {
             $this->setLastError('Impossible de trouver le gestionnaire de tâches');
         } else {
-            $this->verifDroit($daemon->id_e ?? 1, DroitService::getDroitEdition(DroitService::DROIT_DAEMON));
+            $this->verifDroit(
+                $daemon->id_e ?? EntiteSQL::ID_E_ENTITE_RACINE,
+                DroitService::getDroitEdition(DroitService::DROIT_DAEMON)
+            );
             if ($this->getDaemonManager()->status() === DaemonManager::IS_STOPPED) {
-                $this->setLastError('Le gestionnaire tâches global est arrêté');
+                $this->setLastError('Le gestionnaire de tâches global est arrêté');
             } else {
                 $this->getDaemonManager()->startDaemon($id_daemon);
                 $this->setLastMessage('Le gestionnaire de tâches a été lancé sur l\'entité');
@@ -699,14 +747,17 @@ class DaemonControler extends PastellControler
     public function daemonStopAction(): void
     {
         $id_daemon = $this->getGetInfo()->getInt('id_daemon');
-        $return_url = $this->getGetInfo()->get('return_url', 'Daemon/config');
+        $return_url = $this->getGetInfo()->get('return_url', 'Daemon/configuration');
         $daemon = $this->getDaemonSQL()->getDaemon($id_daemon);
         if ($daemon === null) {
             $this->setLastError('Impossible de trouver le gestionnaire de tâches');
         } else {
-            $this->verifDroit($daemon->id_e ?? 1, DroitService::getDroitEdition(DroitService::DROIT_DAEMON));
+            $this->verifDroit(
+                $daemon->id_e ?? EntiteSQL::ID_E_ENTITE_RACINE,
+                DroitService::getDroitEdition(DroitService::DROIT_DAEMON)
+            );
             if ($this->getDaemonManager()->status() === DaemonManager::IS_STOPPED) {
-                $this->setLastError('Le gestionnaire tâches global est arrêté');
+                $this->setLastError('Le gestionnaire de tâches global est arrêté');
             } else {
                 $this->getDaemonManager()->stopDaemon($id_daemon);
                 $this->setLastMessage('Le gestionnaire de tâches a été arrêté sur l\'entité');
@@ -723,12 +774,15 @@ class DaemonControler extends PastellControler
      */
     public function createAction(): void
     {
-        $this->verifDroit(0, 'system:edition');
+        $this->verifDroit(
+            EntiteSQL::ID_E_ENTITE_RACINE,
+            DroitService::getDroitEdition(DroitService::DROIT_DAEMON)
+        );
         $recuperateur = $this->getGetInfo();
         $id_e = $recuperateur->getInt('id_e');
-        if ($id_e === 0 || $this->getDaemonSQL()->getDaemonByEntity($id_e) !== null) {
+        if ($id_e === EntiteSQL::ID_E_ENTITE_RACINE || $this->getDaemonSQL()->getDaemonByEntity($id_e) !== null) {
             $this->setLastError('Un gestionnaire de tâches existe déjà pour cette entité');
-            $this->redirect('Daemon/config');
+            $this->redirect('Daemon/configuration');
         }
         $this->setViewParameter('id_e', $id_e);
         $this->setViewParameter('nb_free_workers', $this->getDaemonSQL()->getNbSharedWorkers() - 1);
@@ -744,19 +798,22 @@ class DaemonControler extends PastellControler
      */
     public function doCreateAction(): void
     {
-        $this->verifDroit(0, 'system:edition');
+        $this->verifDroit(
+            EntiteSQL::ID_E_ENTITE_RACINE,
+            DroitService::getDroitEdition(DroitService::DROIT_DAEMON)
+        );
         $recuperateur = $this->getPostInfo();
         $id_e = $recuperateur->getInt('id_e');
         $nb_allocated_workers = min(
             $recuperateur->getInt('nb_allocated_workers'),
             $this->getDaemonSQL()->getNbSharedWorkers() - 1
         );
-        if ($id_e === 0 || $this->getDaemonSQL()->getDaemonByEntity($id_e) !== null) {
+        if ($id_e === EntiteSQL::ID_E_ENTITE_RACINE || $this->getDaemonSQL()->getDaemonByEntity($id_e) !== null) {
             $this->setLastError('Un gestionnaire de tâches existe déjà pour cette entité');
-            $this->redirect('Daemon/config');
+            $this->redirect('Daemon/configuration');
         }
         $this->getDaemonManager()->addDaemon($id_e, $nb_allocated_workers);
         $this->setLastMessage('Le gestionnaire de tâches a été créé avec succès');
-        $this->redirect("/Daemon/config?id_e=$id_e");
+        $this->redirect("/Daemon/configuration?id_e=$id_e");
     }
 }