+++ controler/PastellControler.php
@@ -70,11 +70,29 @@ class PastellControler extends Controler
      */
     protected function setDroitsDaemon(int $id_e): void
     {
-        $this->setViewParameter('daemon_global_lecture', $this->verifDroit(0, DroitService::getDroitLecture(DroitService::DROIT_DAEMON)));
-        $this->setViewParameter('daemon_lecture', $this->verifDroit($id_e, DroitService::getDroitLecture(DroitService::DROIT_DAEMON)));
-        $this->setViewParameter('daemon_edition', $this->verifDroit($id_e, DroitService::getDroitEdition(DroitService::DROIT_DAEMON)));
+        $this->setViewParameter(
+            'daemon_global_lecture',
+            $this->hasDroit(
+                EntiteSQL::ID_E_ENTITE_RACINE,
+                DroitService::getDroitLecture(DroitService::DROIT_DAEMON)
+            )
+        );
+        $this->setViewParameter(
+            'daemon_lecture',
+            $this->hasDroit($id_e, DroitService::getDroitLecture(DroitService::DROIT_DAEMON))
+        );
+        $this->setViewParameter(
+            'daemon_edition',
+            $this->hasDroit($id_e, DroitService::getDroitEdition(DroitService::DROIT_DAEMON))
+        );
         $this->setViewParameter('daemon_exists', $this->getDaemonSQL()->getDaemonByEntity($id_e));
-        $this->setViewParameter('system_edition', $this->verifDroit(0, DroitService::getDroitEdition(DroitService::DROIT_SYSTEM)));
+        $this->setViewParameter(
+            'system_edition',
+            $this->hasDroit(
+                EntiteSQL::ID_E_ENTITE_RACINE,
+                DroitService::getDroitEdition(DroitService::DROIT_SYSTEM)
+            )
+        );
     }
 
     protected function setDroitImportExportConfig(int $id_e): void
@@ -443,11 +461,6 @@ class PastellControler extends Controler
     public function getDaemonManager(): DaemonManager
     {
         return $this->getInstance(DaemonManager::class);
-    }
-
-    public function getDaemonSQL(): DaemonSQL
-    {
-        return $this->getInstance(DaemonSQL::class);
     }
 
     public function getJobQueueSQL(): JobQueueSQL