+++ template/DaemonConfiguration.php
@@ -3,13 +3,12 @@
 declare(strict_types=1);
 
 /**
- * @var bool $system_edition
  * @var Gabarit $this
  * @var string $search
  * @var int $offset
  * @var array $entity_list
- * @var int $nb_workers
  * @var int $nb_allocated_workers
+ * @var int $nb_workers
  * @var int $nb_shared_workers
  * @var int $global_daemon_status
  */
@@ -46,23 +45,34 @@ declare(strict_types=1);
         </tr>
     </table>
 
-    <?php if ($system_edition) : ?>
-        <a class='btn btn-primary' href="Daemon/editConfig">
-            <i class="fa fa-pencil"></i>&nbsp;
-            Modifier
-        </a>
-    <?php endif ?>
+    <a class='btn btn-primary' href="Daemon/editConfiguration">
+        <i class="fa fa-pencil"></i>&nbsp;
+        Modifier
+    </a>
 </div>
 
 <div class="box">
     <div style='width:100%;'>
         <h2>Liste des gestionnaires de tâches par entité</h2>
     </div>
+    <div class="row">
+        <form action='<?php $this->url('Daemon/configuration') ?>' method='get' class="pt-3 input-group col-md-4">
+            <input type='text' name='search' id='search'
+                   value='<?php echo $search; ?>' class="form-control"
+                   placeholder="Rechercher par dénomination d'entité"/>
+            <button type='submit' class='btn btn-primary btn-search' id="search-entite"><i class='fa fa-search'></i>
+            </button>
+            <div class="col-md-8"></div>
+        </form>
+    </div>
 
-    <?php $this->suivantPrecedent($offset, 20, count($entity_list), "Daemon/config?search=$search"); ?>
-
+    <?php $this->suivantPrecedent($offset, 20, count($entity_list), "Daemon/configuration?search=$search"); ?>
+    <h3 id="title-result" class="ls-off title-result">Résultat(s) de la recherche</h3>
     <form action='<?php $this->url('Daemon/allocate'); ?>' method='post' class="pt-3 input-group col-md-4">
         <?php $this->displayCSRFInput(); ?>
+        <button type="submit" class="btn btn-primary">
+            <i class="fa fa-floppy-o"></i>&nbsp;Enregistrer
+        </button>
         <table class="table table-striped">
             <tr>
                 <th class='w200'>Entité</th>
@@ -88,7 +98,7 @@ declare(strict_types=1);
                         <?php endif; ?>
                     </td>
                     <td>
-                        <?php if ($entity['id_daemon'] !== null && $entity['id_daemon'] !== 1) : ?>
+                        <?php if ($entity['id_daemon'] !== null) : ?>
                             <input type="number" id="daemon_<?= $entity['id_daemon'] ?>"
                                    name="data[<?= $entity['id_daemon'] ?>]" value="<?= $entity['nb_workers'] ?>" min="0"
                                    max="<?= $nb_workers - 1 ?>"/>
@@ -97,11 +107,6 @@ declare(strict_types=1);
                         <?php endif ?>
                     </td>
                     <td>
-                        <?php if ($entity['id_daemon'] !== null && $entity['id_daemon'] !== 1) : ?>
-                            <a class='btn btn-danger'
-                               href='Daemon/deleteDaemon?id_daemon=<?=$entity['id_daemon']?>'
-                            ><i class="fa fa-trash"></i>&nbsp;Supprimer</a>
-                        <?php endif ?>
                         <?php if ($entity['id_daemon'] !== null && $global_daemon_status) : ?>
                             <?php
                             if ($entity['state']) : ?>
@@ -114,6 +119,9 @@ declare(strict_types=1);
                                     <i class="fa fa-play"></i>&nbsp;Démarrer
                                 </a>
                             <?php endif ?>
+                            <a class='btn btn-danger'
+                               href='Daemon/deleteDaemon?id_daemon=<?=$entity['id_daemon']?>'
+                            ><i class="fa fa-trash"></i>&nbsp;Supprimer</a>
                         <?php elseif ($entity['id_daemon'] === null) : ?>
                             <a href="<?= "Daemon/create?id_e={$entity['id_e']}" ?>" class="btn btn-primary"
                                id="arreter_deamon"