+++ template/DaemonDelete.php
@@ -1,5 +1,7 @@
 <?php
 
+declare(strict_types=1);
+
 /** @var Gabarit $this */
 /** @var Daemon $daemon */
 /** @var array $entite */
@@ -13,11 +15,11 @@
         <b><?php hecho($entite['denomination']) ?></b>
     </div>
 
-    <form action='<?php $this->url('Daemon/doDeleteDaemon?id_daemon=' . $daemon->id_daemon); ?>' method='post' >
+    <form action='<?php $this->url("Daemon/doDeleteDaemon?id_daemon=$daemon->id_daemon"); ?>' method='post' >
         <?php $this->displayCSRFInput() ?>
-        <input type='hidden' name='id_t' value='<?php hecho($daemon->id_e)?>' />
+        <input type='hidden' name='id_t' value='<?php hecho((string)$daemon->id_e)?>' />
 
-        <a class='btn btn-outline-primary' href='<?php $this->url('Daemon/config')?>'>
+        <a class='btn btn-outline-primary' href='<?php $this->url('Daemon/configuration')?>'>
             <i class="fa fa-times-circle"></i>&nbsp;Annuler
         </a>
         <button type="submit" class="btn btn-danger">