+++ template/DaemonFrequenceDetail.php
@@ -5,7 +5,7 @@
  * @var Gabarit $this
  */
 ?>
-<a class='btn btn-link' href='<?php $this->url("Daemon/frequenceConfig") ?>'>
+<a class='btn btn-link' href='<?php $this->url("Daemon/frequenceConfiguration") ?>'>
     <i class="fa fa-arrow-left"></i>&nbsp;Retour à la liste des fréquences
 </a>
 <div class="box">